refactor(web-sdk): rename fetch v2 module
